Conflicts:
	Doc/Manual/chapters
	Examples/Makefile.in
	Examples/test-suite/dynamic_cast.i
	Lib/exception.i
	Makefile.in
	Source/Include/swigwarn.h
	Source/Modules/swigmain.cxx
	Source/Swig/swig.h
	Tools/config/config.guess
	Tools/config/config.sub
	configure.in

From: William S Fulton <wsf@fultondesigns.co.uk>

git-svn-id: https://swig.svn.sourceforge.net/svnroot/swig/branches/gsoc2008-maciekd@13036 626c5289-ae23-0410-ae9c-e8d60b6d4f22

+++ Examples/lua/owner/example.cxx
@@ -54,14 +54,14 @@ void ShapeOwner::add(Shape* ptr) // this method takes ownership of the object
 
 Shape* ShapeOwner::get(int idx) // this pointer is still owned by the class (assessor)
 {
-    if (idx<0 || idx>=shapes.size())
+    if (idx < 0 || idx >= static_cast<int>(shapes.size()))
         return NULL;
     return shapes[idx];
 }
 
 Shape* ShapeOwner::remove(int idx) // this method returns memory which must be deleted
 {
-    if (idx<0 || idx>=shapes.size())
+    if (idx < 0 || idx >= static_cast<int>(shapes.size()))
         return NULL;
     Shape* ptr=shapes[idx];
     shapes.erase(shapes.begin()+idx);
